update types and auth handling on layouts
