+++ javascript/todo-list-auth-mongo/server/package.json
@@ -10,7 +10,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@genezio/types": "^1.0.14",
+    "@genezio/types": "^1.0.19",
     "bcryptjs": "^2.4.3",
     "mongoose": "^6.8.3"
   },
