+++ javascript/todo-list-auth-postgres/client/package.json
@@ -11,8 +11,8 @@
     "install-prod-sdk": "npm install @genezio-sdk/todo-list-auth-postgres@1.0.0-prod"
   },
   "dependencies": {
-    "@genezio/auth": "^2.0.19",
-    "@genezio/types": "^1.0.14",
+    "@genezio/auth": "^2.1.1",
+    "@genezio/types": "^1.0.19",
     "bootstrap": "^5.3.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -20,6 +20,7 @@
     "reactstrap": "^9.2.1"
   },
   "devDependencies": {
+    "@genezio/vite-plugin-genezio": "^1.0.5",
     "@types/react": "^18.2.37",
     "@types/react-dom": "^18.2.15",
     "@typescript-eslint/eslint-plugin": "^6.10.0",
@@ -28,8 +29,7 @@
     "eslint": "^8.53.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.4",
-    "vite": "^5.2.9",
-    "@genezio/vite-plugin-genezio": "^1.0.5"
+    "vite": "^5.2.9"
   },
   "prettier": {
     "tabWidth": 2
