+++ javascript/todo-list-auth-postgres/server/package.json
@@ -10,7 +10,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@genezio/types": "^1.0.14",
+    "@genezio/types": "^1.0.19",
     "dotenv": "^16.0.3",
     "pg": "^8.11.3",
     "pg-hstore": "^2.3.4",
