+++ typescript/todo-list-auth-mongo/server/package.json
@@ -10,7 +10,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@genezio/types": "^1.0.14",
+    "@genezio/types": "^1.0.19",
     "bcryptjs": "^2.4.3",
     "dotenv": "^16.0.3",
     "jsonwebtoken": "^8.5.1",
