+++ typescript/todo-list-auth-postgres/client/src/layouts/Admin.tsx
@@ -2,22 +2,18 @@ import { AuthService } from "@genezio/auth";
 import React from "react";
 import { useNavigate } from "react-router-dom";
 
-export default function Auth(props: { element: React.ReactNode }) {
+export default function Admin(props: { element: React.ReactNode }) {
   const navigate = useNavigate();
-
   React.useEffect(() => {
     async function checkUserAuth() {
-      const result = await AuthService.getInstance()
+      await AuthService.getInstance()
         .userInfo()
-        .catch(() => {
-          return null;
+        .catch((error) => {
+          console.log("Error checking user auth", error);
+          navigate("/auth/login");
         });
-      if (result) {
-        navigate("/admin/all-tasks");
-      }
     }
     checkUserAuth();
   }, [navigate]);
-
   return <>{props.element}</>;
 }
