+++ typescript/todo-list-auth-postgres/client/src/views/Register.tsx
@@ -37,7 +37,7 @@ export default function Register() {
       );
       return;
     }
-    navigate("/login");
+    navigate("/auth/login");
   }
 
   return (
@@ -103,7 +103,7 @@ export default function Register() {
                   </div>
                   <div className="mt-2">
                     <span>
-                      Already have an account? <a href="/login">Login</a>
+                      Already have an account? <a href="/auth/login">Login</a>
                     </span>
                   </div>
                 </form>
